backend/requirements.txt
fastapi
uvicorn
openai==2.4.0
python-dotenv
